Some very primitive type discovery, focussed around JVM int type.
git-svn-id: file:///Users/lee/SVNrep/decompiler/trunk@111 c820dd76-da7a-4b58-9bdb-1b1215dfe356

+++ src/org/benf/cfr/reader/entities/ConstantPoolEntryInteger.java
@@ -14,7 +14,7 @@ import org.benf.cfr.reader.util.output.Dumper;
 public class ConstantPoolEntryInteger implements ConstantPoolEntry, ConstantPoolEntryLiteral {
     private final long OFFSET_OF_BYTES = 1;
 
-    private final long value;
+    private final int value;
 
     public ConstantPoolEntryInteger(ByteData data) {
         this.value = data.getS4At(OFFSET_OF_BYTES);
@@ -30,7 +30,7 @@ public class ConstantPoolEntryInteger implements ConstantPoolEntry, ConstantPool
         d.print("CONSTANT_Integer value=" + value);
     }
 
-    public long getValue() {
+    public int getValue() {
         return value;
     }
